+++ src/main/java/com/phantomwing/choppersdelight/datagen/ModBlockTagsProvider.java
@@ -3,6 +3,7 @@ package com.phantomwing.choppersdelight.datagen;
 import com.phantomwing.choppersdelight.ChoppersDelight;
 import com.phantomwing.choppersdelight.Compatibility;
 import com.phantomwing.choppersdelight.block.ModBlocks;
+import com.phantomwing.choppersdelight.tags.CompatibilityTags;
 import com.phantomwing.choppersdelight.tags.ModTags;
 import com.phantomwing.choppersdelight.utils.BlockUtils;
 import net.minecraft.core.HolderLookup;
@@ -54,5 +55,8 @@ public class ModBlockTagsProvider extends BlockTagsProvider {
 
         // Register as mineable with Axe.
         this.tag(BlockTags.MINEABLE_WITH_AXE).addTag(ModTags.Blocks.CUTTING_BOARDS);
+
+        // Register as brittle (Create)
+        this.tag(CompatibilityTags.CREATE_BRITTLE).addTag(ModTags.Blocks.CUTTING_BOARDS);
     }
 }